ptite derniere modif du script

+++ LABORATOIRE/WordPress/4/start_wp.sh
@@ -10,8 +10,8 @@ MYSQL_DATABASE="data_name"
 MYSQL_USER="user_data"
 MYSQL_PASSWORD="mdp"
 
-# Vérifier si le fichier wp-config.php existe déjà pour déterminer si WordPress est installé
-if [ ! -f "$WP_PATH/wp-config.php" ]; then
+# Vérifier si WordPress est installé en vérifiant l'existence d'un fichier essentiel
+if [ ! -f "$WP_PATH/wp-load.php" ]; then
     echo "Installation de WordPress..."
 
     # Créer le répertoire de destination s'il n'existe pas
@@ -22,16 +22,21 @@ if [ ! -f "$WP_PATH/wp-config.php" ]; then
     tar -xzf /tmp/latest.tar.gz -C /tmp
     mv /tmp/wordpress/* "$WP_PATH"
     rm -rf /tmp/latest.tar.gz /tmp/wordpress
+else
+    echo "WordPress est déjà installé."
+fi
 
+# Vérifier si le fichier wp-config.php existe
+if [ ! -f "$WP_PATH/wp-config.php" ]; then
     # Créer le fichier wp-config.php avec WP-CLI
     wp config create --dbname="$MYSQL_DATABASE" --dbuser="$MYSQL_USER" --dbpass="$MYSQL_PASSWORD" --dbhost="localhost" --path="$WP_PATH"
     
     # Installer WordPress avec WP-CLI
     wp core install --url="http://example.com" --title="Mon Site WordPress" --admin_user="admin" --admin_password="admin_password" --admin_email="email@example.com" --path="$WP_PATH"
 
-    echo "WordPress installé avec succès."
+    echo "Fichier wp-config.php créé et WordPress installé avec succès."
 else
-    echo "WordPress est déjà installé."
+    echo "Le fichier wp-config.php existe déjà."
 fi
 
 # Installation et activation des plugins, si nécessaire
@@ -40,3 +45,4 @@ wp redis enable --path="$WP_PATH"
 
 # Démarrer PHP-FPM
 exec php-fpm82 -F
+
